+++ packages/launchpad/src/index.ts
@@ -103,4 +103,4 @@ export { CosmosFeeTable, SigningCosmosClient } from "./signingcosmosclient";
 export { isStdTx, isWrappedStdTx, makeStdTx, CosmosSdkTx, StdTx, WrappedStdTx, WrappedTx } from "./tx";
 export { pubkeyType, PubKey, StdFee, StdSignature } from "./types";
 export { makeCosmoshubPath, executeKdf, KdfConfiguration } from "./wallet";
-export { extractKdfConfiguration, Secp256k1Wallet } from "./secp256k1hdwallet";
+export { extractKdfConfiguration, Secp256k1HdWallet } from "./secp256k1hdwallet";