+++ packages/launchpad/src/secp256k1hdwallet.spec.ts
@@ -3,11 +3,11 @@ import { Secp256k1, Secp256k1Signature, Sha256 } from "@cosmjs/crypto";
 import { fromBase64, fromHex } from "@cosmjs/encoding";
 
 import { serializeSignDoc, StdSignDoc } from "./encoding";
-import { extractKdfConfiguration, Secp256k1Wallet } from "./secp256k1hdwallet";
+import { extractKdfConfiguration, Secp256k1HdWallet } from "./secp256k1hdwallet";
 import { base64Matcher } from "./testutils.spec";
 import { executeKdf, KdfConfiguration } from "./wallet";
 
-describe("Secp256k1Wallet", () => {
+describe("Secp256k1HdWallet", () => {
   // m/44'/118'/0'/0/0
   // pubkey: 02baa4ef93f2ce84592a49b1d729c074eab640112522a7a89f7d03ebab21ded7b6
   const defaultMnemonic = "special sign fit simple patrol salute grocery chicken wheat radar tonight ceiling";
@@ -16,7 +16,7 @@ describe("Secp256k1Wallet", () => {
 
   describe("fromMnemonic", () => {
     it("works", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const wallet = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
       expect(wallet).toBeTruthy();
       expect(wallet.mnemonic).toEqual(defaultMnemonic);
     });
@@ -24,25 +24,25 @@ describe("Secp256k1Wallet", () => {
 
   describe("generate", () => {
     it("defaults to 12 words", async () => {
-      const wallet = await Secp256k1Wallet.generate();
+      const wallet = await Secp256k1HdWallet.generate();
       expect(wallet.mnemonic.split(" ").length).toEqual(12);
     });
 
     it("can use different mnemonic lengths", async () => {
-      expect((await Secp256k1Wallet.generate(12)).mnemonic.split(" ").length).toEqual(12);
-      expect((await Secp256k1Wallet.generate(15)).mnemonic.split(" ").length).toEqual(15);
-      expect((await Secp256k1Wallet.generate(18)).mnemonic.split(" ").length).toEqual(18);
-      expect((await Secp256k1Wallet.generate(21)).mnemonic.split(" ").length).toEqual(21);
-      expect((await Secp256k1Wallet.generate(24)).mnemonic.split(" ").length).toEqual(24);
+      expect((await Secp256k1HdWallet.generate(12)).mnemonic.split(" ").length).toEqual(12);
+      expect((await Secp256k1HdWallet.generate(15)).mnemonic.split(" ").length).toEqual(15);
+      expect((await Secp256k1HdWallet.generate(18)).mnemonic.split(" ").length).toEqual(18);
+      expect((await Secp256k1HdWallet.generate(21)).mnemonic.split(" ").length).toEqual(21);
+      expect((await Secp256k1HdWallet.generate(24)).mnemonic.split(" ").length).toEqual(24);
     });
   });
 
   describe("deserialize", () => {
     it("can restore", async () => {
-      const original = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const original = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
       const password = "123";
       const serialized = await original.serialize(password);
-      const deserialized = await Secp256k1Wallet.deserialize(serialized, password);
+      const deserialized = await Secp256k1HdWallet.deserialize(serialized, password);
       expect(deserialized.mnemonic).toEqual(defaultMnemonic);
       expect(await deserialized.getAccounts()).toEqual([
         {
@@ -59,7 +59,7 @@ describe("Secp256k1Wallet", () => {
       const password = "123";
       let serialized: string;
       {
-        const original = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+        const original = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
         const anyKdfParams: KdfConfiguration = {
           algorithm: "argon2id",
           params: {
@@ -75,7 +75,7 @@ describe("Secp256k1Wallet", () => {
       {
         const kdfConfiguration = extractKdfConfiguration(serialized);
         const encryptionKey = await executeKdf(password, kdfConfiguration);
-        const deserialized = await Secp256k1Wallet.deserializeWithEncryptionKey(serialized, encryptionKey);
+        const deserialized = await Secp256k1HdWallet.deserializeWithEncryptionKey(serialized, encryptionKey);
         expect(deserialized.mnemonic).toEqual(defaultMnemonic);
         expect(await deserialized.getAccounts()).toEqual([
           {
@@ -90,7 +90,7 @@ describe("Secp256k1Wallet", () => {
 
   describe("getAccounts", () => {
     it("resolves to a list of accounts", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const wallet = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
       const accounts = await wallet.getAccounts();
       expect(accounts.length).toEqual(1);
       expect(accounts[0]).toEqual({
@@ -101,7 +101,7 @@ describe("Secp256k1Wallet", () => {
     });
 
     it("creates the same address as Go implementation", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(
+      const wallet = await Secp256k1HdWallet.fromMnemonic(
         "oyster design unusual machine spread century engine gravity focus cave carry slot",
       );
       const [{ address }] = await wallet.getAccounts();
@@ -111,7 +111,7 @@ describe("Secp256k1Wallet", () => {
 
   describe("sign", () => {
     it("resolves to valid signature if enabled", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const wallet = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
       const signDoc: StdSignDoc = {
         msgs: [],
         fee: { amount: [], gas: "23" },
@@ -133,7 +133,7 @@ describe("Secp256k1Wallet", () => {
 
   describe("serialize", () => {
     it("can save with password", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const wallet = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
       const serialized = await wallet.serialize("123");
       expect(JSON.parse(serialized)).toEqual({
         type: "secp256k1wallet-v1",
@@ -155,7 +155,7 @@ describe("Secp256k1Wallet", () => {
 
   describe("serializeWithEncryptionKey", () => {
     it("can save with password", async () => {
-      const wallet = await Secp256k1Wallet.fromMnemonic(defaultMnemonic);
+      const wallet = await Secp256k1HdWallet.fromMnemonic(defaultMnemonic);
 
       const key = fromHex("aabb221100aabb332211aabb33221100aabb221100aabb332211aabb33221100");
       const customKdfConfiguration: KdfConfiguration = {