+++ scripts/wasmd/send_first.js
@@ -5,7 +5,7 @@ const { Random } = require("@cosmjs/crypto");
 const { Bech32 } = require("@cosmjs/encoding");
 const {
   coins,
-  Secp256k1Wallet,
+  Secp256k1HdWallet,
   SigningCosmosClient,
   assertIsBroadcastTxSuccess,
 } = require("@cosmjs/launchpad");
@@ -18,7 +18,7 @@ const faucet = {
 };
 
 async function main() {
-  const wallet = await Secp256k1Wallet.fromMnemonic(faucet.mnemonic);
+  const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
   const client = new SigningCosmosClient(httpUrl, faucet.address0, wallet);
   const recipient = Bech32.encode("cosmos", Random.getBytes(20));
   const amount = coins(226644, "ucosm");