+++ CHANGELOG.md
@@ -59,6 +59,8 @@
   better represent the Amino tx interface. Deprecate `CosmosSdkTx`, which is an
   alias for `WrappedStdTx`.
 - @cosmjs/launchpad: Add `makeStdTx` to create an `StdTx`.
+- @cosmjs/launchpad: Rename `Secp256k1Wallet` to `Secp256k1HdWallet`. Later on,
+  we'll use `Secp256k1Wallet` for single key wallets.
 - @cosmjs/launchpad-ledger: Add package supporting Ledger device integration for
   Launchpad. Two new classes are provided: `LedgerSigner` (for most use cases)
   and `LaunchpadLedger` for more fine-grained access.